fix: rename component
